"Explicit markup ends without a blank line; unexpected unindent"
This one is surprisingly complicated.

Some of the reference documentation does this:

```
Examples
--------
..
>>> foo()
```

This warning is about the `..` on its own line.

Although `..`s like that have been there since 2016 (e.g. 9c7af54de5ce096479a018435e1d514537b7c80e), I'm pretty sure they're syntax errors, and also pointless?

In reStructuredText, `..` is the beginning of an [explicit markup block][1].  But it usually (always?) is followed by the name of a directive, e.g. `.. hint::`.  It's also supposed to be followed by an indented block, which is terminated by a blank line.  None of that happens here.

Lines beginning with `>>>` are [doctest blocks][2], which Sphinx picks up independently.  They're unaffected by the preceding `..`.

(Also, the "Examples" section is treated specially by [the Napoleon extension][3].  But I don't think that has any bearing here.)

So, we just remove all the errant `..` lines.

[1]: https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#explicit-markup-blocks
[2]: https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#doctest-blocks
[3]: https://www.sphinx-doc.org/en/master/usage/extensions/napoleon.html