Fix syntax of "new in version" markers for params
The .. versionadded markers in these param lists had wrong indentation.
In one case, this caused the list to be split up when it was rendered.

Fix this by using a method-level `.. versionadded` instead of param-level ones.

We could also have fixed the indentation to get the param-level markers to work.  But, that would have rendered with slightly uglier formatting.  See sphinx-doc/sphinx#4227.